typedef/refactor: schema for coordinates. Env schema: new variable ENCRYPTION_SECRET. Env schema now has its own file
