+++ .env.example
@@ -7,6 +7,7 @@ ACCESS_TOKEN_SECRET = 379f97530d3b0188aeacfbfd6478d2786d43a160385f07dfab4b687c51
 REFRESH_TOKEN_SECRET = 66b8833ac2d938d066916c238b88c74dfd706c996904cf507ef3c7ee5535403b
 ACCESS_TOKEN_EXPIRY = "30s"
 REFRESH_TOKEN_EXPIRY = "60s"
+ENCRYPTION_SECRET = "017828df74f447a982a889ff3c8476ec87aa15073d0a0b1b34d87cb72df896a6"
